improvement: changing files location
