+++ server/src/controllers/index.ts
@@ -1,2 +1,4 @@
 export * from "./github-auth";
-export * from "./talk";
+export * from "./talk/update-talk";
+export * from "./talk/create-talk";
+export * from "./talk/get-top-talks";
